pattern-file-duplicacy-fix
